Port v5.1 New Compression "Move the shared code out of page_zip_decompress_* functions"
Summary:
Decompressing a page consists of the following steps:
1 - Restore the records from the uncompressed page image.
2 - Clear the unused heap space in the uncompressed page.
3 - Apply modification log.
4 - Verify that the trailer part of the page did not
overlap with the data part of the page.
5 - Restore the uncompressed fields (trx_id, roll_ptr for
primary leaf pages, REC_NODE_PTR length fields for node pointer
pages, nothing for secondary leaf pages.)
6 - Set extra bytes for each record.

Except for the first step, all of the steps are almost identical
for all page types. In addition to this, when we serialize the
records on the page before compression, we will do this in the
first step and not change the other steps. This revision takes the
steps 2 - 6 outside of page_zip_decompress_* functions.

Internal:

This is a direct port D606781, except minor code-formatting/indentation
changes.

Test Plan: Jenkins (arc unit)

Reviewers: rongrong, nizamordulu

Reviewed By: tianx
(Original Reviewed By: nizamordulu)

CC: jtolmer, nizamordulu, mysql-eng@

Differential Revision: https://phabricator.fb.com/D853501

Task ID: 2518197