Port v5.1 Compression: Factor out record serialization.
Summary:
Factor out serialization of a record from page_zip_write_rec().

The main issue with serializing the records is that normally
the offsets are stored backwards and this causes problems during
deserialization. The code that writes to the modification log already
reverses the direction of the offsets to be able to apply the
modification log when the page is decompressed. We use the same code to
create a function to serialize a record to a data stream. This function
will be reused for serializing the records before compressing the page.

Test Plan: Jenkins (arc unit)

Reviewers: rongrong, nizamordulu

Reviewed By: nizamordulu