Fix rebase error in per-page-type table stats.
Summary:
During a rebase to a new upstream mysql version, this was fixed wrong.

Stat gen was linked to the wrong variable.  The "slot" variable was the
index of an earlier loop, not current slot (which is "aio_slot").

Internal:
This should be squashed with:
	https://phabricator.fb.com/D654945
	https://phabricator.fb.com/D774310

Test Plan:
This was found with the new valgrind tests in Jenkins.

This was also found by the included mtr test, but this was fixed wrong.
The results file was updated with the incorrect results, so it passed.

Reviewers: rongrong

Reviewed By: rongrong

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D867632

Task ID: 2530404
(cherry picked from commit bf840df09401ade71b3b3b7cb2a5741efb11d85c)