Fix rebase error in per-page-type table stats.
Summary:
During a rebase to a new upstream mysql version, this was fixed wrong.

Stat gen was linked to the wrong variable.  The "slot" variable was the
index of an earlier loop, not current slot (which is "aio_slot").

Test Plan:
This was found with the new valgrind tests in Jenkins.

This was also found by the included mtr test, but this was fixed wrong.
The results file was updated with the incorrect results, so it passed.

Reviewers: rongrong

Reviewed By: rongrong