Fix for MySQL BUG#69471
Summary:
Fix for MySQL BUG#69471: UNION of derived tables returns wrong results with false clauses

There already existed some logic so that if we are optimizing
SELECT2 in SELECT1 ... FROM (SELECT2) we never set best_rowcount
to zero or one unless we are sure that this count is exact.
Indeed a count of zero or one has the consequence that SELECT1
treats the derived table as constant, a decision which cannot be rolled
back, and which leads to catastrophic consequences if, when we later
materialize, the derived table turns out to not have the
anticipated row count. The logic in question forgot to apply
itself to this case: SELECT1 ... FROM (SELECT2 UNION SELECT3 etc.)

We got this patch from ORACLE.

Internal:

This is a direct port of commit130627-4UOL[9], which we got from Oracle.

Test Plan:
Tested

CREATE TABLE table1 ( a INT, b INT);
CREATE TABLE table2 ( a INT, b INT);
INSERT INTO table1 VALUES (1, 4);
INSERT INTO table1 VALUES (2, 4);
INSERT INTO table2 VALUES (1, 5);
INSERT INTO table2 VALUES (2, 5);

SELECT *
FROM  (
	(SELECT *
           FROM table1
          WHERE 1=0)
UNION
       	 (SELECT *
     	    FROM (SELECT * from table2) t2
           WHERE t2.b = 5)
        ) a3
;

on new 5.6 RPM.

Returns the result

1	5
2	5

correctly now.

Reviewers: jtolmer, steaphan

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D871217

Task ID: 2520695