Try to fix failing rpl.rpl_semi_sync_group_commit_deadlock
Summary:
The test is often failing under valgrind with a warning printed to the
.err log that semi-sync timedout waiting for a reply from the slave. Looking
at the test shows that the timeout is set to a very large value for most of
the test, so it shouldn't be hitting a timeout. My theory is that the warning
happens when the master is waiting for a semi-sync ACK at the point when the
test does:
  --source include/uninstall_semisync.inc

My attempt at fixing is to synchronize the master and slave prior to uninstalling
semisync so that there shouldn't be any waiting transactions.

Test Plan: Jenkins

Reviewers: steaphan

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D876685

Task ID: 2566063