Add a lint engine for MySQL 5.6 Repo
Summary:
@internal

This diff adds a lint engine for MySQL 5.6 repo. The linter notifies
about the basic syntax errors:

1. DOS newlines instead of UNIX newlines
2. Lines exceeding 80 characters
3. Files not ending in new line
4. Lines containing trailing whitespace

The linter notifies only about the lines that have been modified, except
result files, and documentation. In other words, it keeps all the bad
formatting from upstream.

The linter also provides an option to auto-fix, when a
modified line contains trailing whitespace.

The linter assumes that an innodb tab is 8 spaces in width.

I had looked at the documentation of configuring a new project (http://www.phabricator.com/docs/phabricator/article/Arcanist_User_Guide_Configuring_a_New_Project.html) and at a lot of examples and source code on the internet (mostly at the source code of these classes).

Let me write an explanation of what this code does:

In .arcconfig, you just specify the id of the lint engine (which in our case is FacebookMysqlLintEngine).

In the phutil_map file, you specify the path to the lint engine, and path to all the custom linters that you use, and also (id, base_class) for each custom linter (You always create a linter that derives itself from ArcanistLinter - look at fbcode's linter for example).

In FacebookMysqlLintEngine.php, you override the abstract function buildLinters() to return the array of linters (the ones you have created, and the existing ones you wish to use) you'd like to use for your project. We use ArcanistGeneratedLinter, ArcanistFilenameLinter, ArcanistNoLintLinter, and ArcanistSpellingLinter (read more on internet) from the existing ones, and we create one of our own: FacebookMysqlLinter. I also ran a script to get the different extensions we have in our repo, and added the ones that I thought were fit for linting.

In FacebookMysqlLinter.php, we define our linter. I looked mostly at ArcanistTextLinter to write code for this. It derives from ArcanistLinter and specifies the various checks that we need in lintPath by overriding it. Using ArcanistLintSeverity::SEVERITY_WARNING, we ensure that the warnings are raised only for the code that we modified. If we instead used ArcanistLintSeverity::SEVERITY_ERROR, it would raise a lint error at every place where there is one, whether or not it came from us, or upstream. Using ArcanistLintSeverity::SEVERITY_AUTOFIX for trailing whitespace, we give an option to user to apply the patch (to get rid of trailing whitespace) that Arcanist creates.

This diff adds some basic indentation error finding support for arc
lint. Specifically, it notifies the user if the indentation level is at
least how much it should actually be. It may not notify the user if the
indentation level is more. I'll add more support in incremental diffs.
...except this is disabled, because that shit was all broken.

Test Plan:
1. arc unit (Jenkins)

2. Modify an existing file to see that the linter reports the warnings
correctly.

   Warning  (MYSQL2) Line Too Long
    This line is 83 characters long, but the convention is 80
characters.

             950 	}
             951 }
             952
    >>>      953
    >>>      ///////////////////////////////////////////////////////////////////////////////////
             954
             955
/**********************************************************************//**
             956 Restore uncompressed fields (trx id, rollback ptr,
blobs) of all of the

   Auto-Fix  (MYSQL4) Trailing Whitespace
    This line contains trailing whitespace.

             956 Restore uncompressed fields (trx id, rollback ptr,
blobs) of all of the
             957 records on a page from the trailer of the compressed
page */
             958 static
    >>> -    959 void
        +        void
             960 page_zip_restore_uncompressed_fields_all(
             961 	page_zip_des_t*	page_zip,	/*!< in:
page_zip->data trailer and
             962 					page header.
out: page_zip->n_blobs; */
--- /data/users/rahulgulati/mysql/5.6/storage/innobase/page/page0zip.cc
2013-07-11 03:29:24.000000000 -0700
+++ /tmp/epg4b8ahdp4ckwk0/22598-1ePNVx	2013-07-11 04:06:25.000000000
-0700
@@ -956,7 +956,7 @@
 Restore uncompressed fields (trx id, rollback ptr, blobs) of all of the
 records on a page from the trailer of the compressed page */
 static
-void
+void
 page_zip_restore_uncompressed_fields_all(
 	page_zip_des_t*	page_zip,	/*!< in: page_zip->data trailer
and
 					page header. out:
page_zip->n_blobs; */

    Apply this patch to storage/innobase/page/page0zip.cc? [y/N]

Do 'arc lint' and see the following lint errors:

   Warning  (MYSQL_CPP1) Indentation
    This line possibly has an indentation error

              90 UNIV_INTERN my_bool page_zip_debug = FALSE;
              91
              92 struct hello {
    >>>       93 int x;
              94 	int y;
              95 };
              96

   Warning  (MYSQL_CPP1) Indentation
    This line possibly has an indentation error

              96
              97 for (int i = 0; i < N; ++i) {
              98 	int x;
    >>>       99 int y;
             100 }
             101
             102 if (expr) {

   Warning  (MYSQL_CPP1) Indentation
    This line possibly has an indentation error

             101
             102 if (expr) {
             103 	int a;
    >>>      104 int b;
             105 	while (expr) {
             106 		int c;
             107 	int d;

   Warning  (MYSQL_CPP1) Indentation
    This line possibly has an indentation error

             104 int b;
             105 	while (expr) {
             106 		int c;
    >>>      107 	int d;
             108 		for (expr; expr; expr) {
             109 			int e;
             110 		int f;

   Warning  (MYSQL_CPP1) Indentation
    This line possibly has an indentation error

             107 	int d;
             108 		for (expr; expr; expr) {
             109 			int e;
    >>>      110 		int f;
             111 		}
             112 	}
             113 }

I also tested this on a lot of our diffs. There were no false positives and it was able to identify a few indentation errors that got by.

Reviewers: jtolmer, steaphan, nizamordulu

Reviewed By: steaphan, jtolmer

CC: santoshb, mysql-eng@

Differential Revision: https://phabricator.fb.com/D882454
Differential Revision: https://phabricator.fb.com/D883950
Differential Revision: https://phabricator.fb.com/D885949
Differential Revision: https://phabricator.fb.com/D891272
Differential Revision: https://phabricator.fb.com/D910008
Differential Revision: https://phabricator.fb.com/D923046
Differential Revision: https://phabricator.fb.com/D933202
Differential Revision: https://phabricator.fb.com/D960558
Differential Revision: https://phabricator.fb.com/D1053657

Task ID: 2599330, 2604277, 2628376, 2679591, 2603785, 2753518, 2845228, 2756819