Modify severity of ArcanistSpellingLinter fields
Summary:
This diff modifies severity of LINT_SPELLING_PICKY and
LINT_SPELLING_IMPORTANT so that they only show spelling mistakes in
modified lines of code. Also, added a comment for each of the
Linters that we use, and added another extension.

Internal:

Squash with D882454 if we release this to public.

Test Plan: 'arc lint' doesn't produce spelling errors on upstream code now.

Reviewers: steaphan

Reviewed By: steaphan

CC: jtolmer, santoshb, mysql-eng@

Differential Revision: https://phabricator.fb.com/D883950

Task ID: 2604277