Write missing binlog events in case of a slave crash.
Summary:
A slave running with sync_binlog=0 may have missing binlog events in case
of a crash, but the events are committed in the storage engine. With new gtid crash
safe functionality using a transaction table slave_gtid_info, the missing events
can be detected. This diff adds the functionality to write the missing events into
the binlog, only when using statement based replication.

Internal: Should be squashed with https://phabricator.fb.com/D852204.

Test Plan:
verified the output of show binlog events to find the missing events
on a crashed slave. Simulated missing events on a slave in a mtr_test
and re-executed all the events of the slave binlog after resetting
the slave. Finally a consistency check of all the tables between master
and slave is done to verify the functionality.

Reviewers: jtolmer, rudradevbasak

Reviewed By: rudradevbasak

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D888622

Task ID: 2414895