Remove mysql client warning about passwords
Summary:
These warnings were added in 5.6. We get rid of these warnings to make
the functionality as close as to that of 5.1. We'll revert this diff,
and these warnings back again later.

Test Plan:
arc unit (Jenkins)

Do a 'git grep -n "Warning: Using a password on the command line
interface can be insecure."' and re-run all the tests with --record.

Reviewers: jtolmer, chip, steaphan

Reviewed By: steaphan