Adding Histogram to track Binlog Fsyncs
Summary:
This diff adds a histogram (histogram_binlog_fsync) to track the latencies caused by binlog fsyncs. Along with this addition, this diff also adds a system variable which is responsible for configuring the step size of this histogram, similar to previous histograms.
The system variable added is histogram_step_size_binlog_fsync
It is a string of the form "<Value><us/ms/s>". The <Value> is of double data type.

Test Plan:
histogram_step_size_binlog_fsync_basic
show_variables_histogram
jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong