Display Histograms in SHOW GLOBAL STATUS
Summary:
This diff prints all the histogram data to the output of SHOW GLOBAL STATUS.
The value corrersponding to each bucket (of all the histograms) is displayed, and the bucket range (that is to be displayed in microseconds/milliseconds/seconds) is calculated on the fly.

Following is the format in which Histogram data is displayed in the output of SHOW GLOBAL STATUS:
<HistogramName>_<BucketLowerValue>-<BucketUpperValue><Unit>

Presently all the numeric values in the test output are replaced by constants, because after each run the output will not match with the previous.
In future a rigorous test can be written which validates the histogram values.

Test Plan:
innodb_histogram_show_global_status
jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong