Display Histograms in SHOW ENGINE INNODB STATUS
Summary:
This diff displays the Histogram data in the output of SHOW ENGINE INNODB STATUS
Following is the format in which the data is displayed:
Histograms

Async Reads
Bucket xxx to xxx units 	Value
Bucket xxx to xxx units		Value
...............

Async Writes
.....................

and so on..

In the test output, I replaced all numeric values string constant, since these cannot be the same on each run(same is the case with the histogram bin range, because of them being of double data type).
However under the section "File I/O", the state of the Innodb threads is changing frequently on each run. While some are waiting, the others have completed and so on. So I cannot compare the output of SHOW ENGINE INNODB STATUS with the output of a previous run in any case.

A more complicated test cam be written in the future to extract certain output using Perl and then perform validation.

I checked by manually running the command and the histogram is getting printed in the correct format.It can be easily reproduced on any machine

Test Plan:
innodb_histogram_show_innodb_status
jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D891418

Task ID: 1402576