Modify Linter so that it correctly calculates the tab widths
Summary:
Earlier, the linter used to consider the width of a tab as 8
characters, which is the indentation that is used by MySQL, but the next
position that a tab takes you at is the next multiple of 8. We modify
the linter to follow this convention.

Internal:

Squash with D882454

Test Plan:
Tested this approach for lines < 80 characters, exactly 80
characters, and 81 and > 81 characters. The lint warnings are only shown
for lines that have 81 or more characters.

Reviewers: steaphan

Reviewed By: steaphan

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D891272

Task ID: 2628376