Adding Histogram for Connection Create Latency per User
Summary:
This diff adds a new histogram (histogram_connection_create) in order to track the connection create latency per user. It also adds a configurable system variable to set the step size of this histogram. This variable is named histogram_step_size_connection_create.
(The configurable step size variable is common for all users)
I have assumed that the time taken for a connection to be created is a sum of the following times:
1. Time taken to launch the new thread (which will handle the connection)
2. Time taken to login the connection and authenticate the user using Access Control Lists etc
3. Time taken to prepare connection state

This way we will be able to track the total time taken starting from when a new thread is spawned upto the time the connection is ready to handle queries.

Test Plan:
show_variables_histogram
histogram_step_size_connection_create_basic
jenkins

Reviewers: steaphan, rongrong

Reviewed By: steaphan