Port v5.1 "Factor out deserialization of a record" to 5.6
Summary:
Port v5.1 "Factor out deserialization of a record from
page_zip_apply_log()" to 5.6

When the modification log is applied to the uncompressed version of the
compressed page image, the records are read from the modification log.
We factor out this piece of code into its own function so that we can
use this deserialization function when we need to deserialize the
serialization of records on a page.

Internal:

This is a direct port of D608326 except minor
indentation/formatting/spelling changes.

Test Plan: arc unit

Reviewers: nizamordulu

Reviewed By: tianx
(Original Reviewed By: nizamordulu)

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D896448

Task ID: 2528687