Fix the bugs introduced by port of "Improve page_zip_validate()"
Summary:

This ported some code to improve the functionality of page_zip_validate(),
but similar improvements were already made to page_zip_validate() in 5.6,
and some of these conflicted.  This did not merge them properly.

Another bug that existed prior to this was that a heap object was not
always freed after being used in page_zip_validate_low().
This diff fixes this by always freeing the heap object.

Test Plan:
Build with UNIV_ZIP_DEBUG enabled and run innodb_stress.

Reviewers: rahulgulati

Reviewed By: rahulgulati