Fixes for latency histograms
Summary:
This includes two small fixes for the previous latency histogram commits.

A small fix which prevents writing of a histogram value outside the
maximum bucket range.  This involves subtracting a 1 in the formula for
searching a value in the histogram.  This was causing assertion failures
for me when I was adding Histograms for per Statement Latencies per
User.  Santosh helped me debug this.

A fix to a bug which I encountered while testing of the values of
histogram buckets, and if they actually sum up to the number of requests
of that corresponding statistic.  The bug was that there is an increment
to histogram_async_read when actually there should be an increment to
histogram_async_write.  Fixed this here.

Test Plan: Jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong