Fix for Diff D885994
Summary:
This diff is small fix which prevents writing of a histogram value outside the maximum bucket range. It involves subtracting a 1 in the formula for searching a value in the histogram.
This was causing assertion failures for me when I was adding Histograms for per Statement Latencies per User. Santosh helped me debug this.

Internal: This is a fix to D885994, which is a replacement diff for D866912 and D878812

Test Plan: jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D903046

Blame Revision: D885994

Task ID: 1402576