Test for validating Latency Histogram bucket values in SHOW ENGINE INNODB STATUS
Summary:
This diff adds rigorous testing for checking if the histogram bucket values , that are replaced with a constant because they change on each run, are actually valid.
The testing is done as follows:
In the output of SHOW ENGINE INNODB STATUS, there are statistics printed corresponding to Async read|writes, Sync reads|writes and so on.
The test processes the output of SHOW ENGINE INNODB STATUS to extract the number of requests (for example, X) in :
Async reads: X requests, ....

It then extracts the bucket values of that histogram(for example here, histogram_async_read), before replacing them with a string constant, and sums all of them up and checks whether they are equal to X.

Note: SHOW ENGINE INNODB STATUS prints the number of fsyncs in front of the statistic 'File Flushes:'. (I checked this in the file os0file.cc). I added two separate histograms for file flush time and fsync time, where the former might contain multiple calls to fsync if there is no lock available.
So I could not yet figure out how to test the values of histogram_file_flush_time.

Test Plan:
innodb_histogram_show_innodb_status
jenkins

Reviewers: rongrong, steaphan

Reviewed By: steaphan

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D905667

Task ID: 1402576