Fix for Diff D885994
Summary:
This is a fix to a bug in D885994 (which further is a replacement for D866912 and D878812) which I encountered while testing of the values of histogram buckets,and if they actually sum up to the number of requests of that corresponding statistic.
The bug was that there is an increment to histogram_async_read when actually there should be an increment to histogram_async_write. Fixed this here.

Internal: This is fix for D885994. It should be squashed with this diff.

Test Plan: jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D905411

Blame Revision: D885994

Task ID: 1402576