Test for validating Latency Histogram bucket values in Information_Schema.User_Latency_Histograms table
Summary:
This diff adds a test which is used to validate the histogram values corresponding to the per Statement type latencies. It works as follows:
The perl script sums up all the values of histogram buckets for each statement type, and then compares it to the commands_xxx value in the user_statistics table, corresponding to that statement.
For example: if the statement type is INSERT,
the script retrieves the row with type='INSERT' from information_schema.user_latency_histograms and sums up all the columns values. It then compares this sum to the 'commands_insert' value in information_schema.user_latency_histograms
If the values match, it prints out a message.

Test Plan:
user_latency_histograms
jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D906471

Task ID: 1402576