Fix intermittent drop table valgrind failure
Summary:
When the page_cleaner thread tries to flush pages in the neighbourhood
of the page in consideration, it release the buffer pool mutex, which
the foreground drop table thread may acquire, and then delete the
tablespace associated with the page. In such a scenario, space_size
would receive a value of zero, and no flushing would happen. This is
totally fine since that would only happen when a table is dropped, and
we wouldn't want to flush pages in that case anyway.

Test Plan: arc unit (Jenkins)

Reviewers: jtolmer, nizamordulu

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D906058

Task ID: 2628489