Improve Lint performance by avoid calling getData() again and again
Summary:
The callback lintPath() in FacebookMysqlLinter.php further calls lint
functions, each of which calls getData($path) which returns the contents
of the file at 'path'. Since the data isn't modified by these functions,
we can just store the data in a class variable, initialized in the
callback itself, and avoid calling getData() again and again, and
instead use the variable.

Since the variable data is used by the base class ArcanistLinter, we
store the data in a variable fileData.

Internal: Squash with D882454 as the improvements aren't significant enough that we have a separate diff.

Test Plan: 1. Test that all the lint features still work.

Reviewers: steaphan

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D910008

Task ID: 2679591