Fix flushed_by stats in mysql-5.6
Summary:
the buffer pool flushed by stats in 5.6 are not collected properly.
1. Innodb_buffer_pool_pages_flushed only tracks lru and list, not single_page.
2. Innodb_buffer_pool_pages_flushed_by_lru, Innodb_buffer_pool_pages_flushed_by_list,
   and Innodb_buffer_pool_pages_flushed_by_single_page are not populated.

Test Plan:
A new test, innodb.innodb_buffer_pool_flushed_stats is added.

This new test ensures that the sum of the by-type stats roughly equals
the total stat.  Unfortunately, the total stat is a fuzzy counter which
has no thread protection and isn't guaranteed to be 100% accurate. So,
it does only an in-bounds compare, not an exact one.

Reviewers: steaphan

Reviewed By: steaphan