More refactoring of the connect path; introduce a state machine
Summary:
More preparation for the async client.  This splits the connect
codepaths further into smaller units (generally separating at the IO
boundaries).  It changes the mechanism for executing the connect
sequence to mimic what the async version will do.

This doesn't break apart the auth code yet, but it will have a similar
splitting.  Once that's done, then most of this code will be ready to be
used in both sync and async modes.

Test Plan: mtr

Reviewers: santoshb

Reviewed By: santoshb

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D927744