Fix MySQL Bug# 70026
Summary:
This is a fix for MySQL Bug# 70026 (Auto reconnect does not work with
5.6 libmysqlclient). This change was introduced by revno 3133, authored
by Davi Arnaut. He has provided a patch for the bug at
https://gist.github.com/darnaut/6246624. This is a direct port of the
patch provided by him.

Test Plan:
Test that without the patch, the connection to mysql is lost, and with
the patch, the re-connection works fine.

Reviewers: steaphan, santoshb

Reviewed By: santoshb

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D930851

Task ID: 2232993