Treat missing server lra support only as a warning in mysqldump
Summary:
Specifying --lra* options to mysqldump on a server without the actual
logical read ahead support would cause an error.

This makes it into a warning instead.

If the initial --lra-size SET succeeds, and somehow one of the other
SETs fails, this is still an error (something is very wrong).

Test Plan:
Ran mysqldump --lra-size=8 on a table on 5.1 and 5.6 servers.
On 5.1, printed the warning, and dumped normally.
On 5.6, no warning, and dumped normally.

Reviewers: santoshb, nizamordulu

Reviewed By: nizamordulu