Make innodb_doublewrite test more robust against valgrind issues.
Summary:
This test relied on the timing of mysqld to flush before the server
was shut down. Apparently when run under valgrind, this was not always
the case. In order to make sure that the table pages are flushed while
the server is being shut down, we send the "FLUSH TABLES" command over
a separate connection and don't wait for that command to finish before
shutting down the server.

Internal:
Squash this with D922787.

Test Plan: Run innodb_doublewrite under valgrind.

Reviewers: jtolmer, steaphan

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D934695

Task ID: 2747799