Do not call close() on an invalid socket when checking for IPv6.
Summary:
When mysqld initializes the network subsystem it checks for
IPv6 support by first trying to create an IPv6 socket then
checking if an IPv6 socket can be resolved. If the system
doesn't have IPv6 support the socket created is -1 but mysqld
still calls close() on this invalid descriptor. This causes
warnings when mysqld is run under valgrind.

Test Plan: Run mtr tests with valgrind.

Reviewers: jtolmer, santoshb

Reviewed By: santoshb