fix missing big transaction with GTIDs
Summary:
This diff ports Oracles's patch for missing big transaction
with GTIDs. This diff also includes fix for missing big transaction
with GTIDs when parallel replication is enabled.

Io_thread may receive only a partial transaction before
it was stopped using stop slave. This causes a partial transaction
with GTID to get logged in the relay log. When the slave is restarted
again, it misses the transaction because GTID protocol assumes that the logged
GTID in relay log is complete. This is fixed by removing the last GTID
in the relay log from the gtid_retrieved_set causing the master to resend that
whole transaction.

Possible cases:
1) If there is a partial transaction, the whole transaction is retrieved
again into the next relay log which will be executed by SQL thread. SQL thread
rollbacks the partial transaction after seeing the FDE in the next relay log
and starts executing the same transaction which was retrieved again. In MTS
mode SQL thread appends a ROLLBACK query to the slave worker queue which got
the partial transaction.

2) I/O thread would have retrieved full transaction already and SQL thread
would have already executed it. In that case, We are not going to remove last
retrieved GTID from "Retrieved_gtid_set" otherwise we will see gaps in
"Retrieved set".

3) I/O thread would have retrieved full transaction already in the first time
itself and SQL thread has not applied it yet while requesting dump but applied
it after I/O thread started receiving events from master. In this case
retrieving the same transaction again will not cause problem because GTID
number is same, Hence SQL thread will not commit it again.

Please note there will be partial transactions written in relay log but they
will not cause any problem in case of transactional tables.  But in case
of non-transaction tables, partial transaction will create inconsistency
between master and slave.  In that case, users need to check manually.
This is not a problem for us since we are using transactional tables.

Test Plan:
Added a test to verify all the scenarios with and without MTS. Also
ran mysqltest.sh --parallel=32 with and without valgrind.

Reviewers: rudradevbasak, jtolmer, yoshinori

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D938963

Task ID: 2701928
(cherry picked from commit 7013f0b166ffdad6fed172f17f1f67dde4ded6a2)