Clean up libreadline build fix.
Summary:
I left this in a state which required options be passed to build.

Now, it doesn't, it always builds with libreadline.

Internal:
I will squash this with: https://phabricator.fb.com/D932275

Test Plan:
Built it with our tools (with the options removed), and with raw cmake.
And I confirmed the client was still linked to libreadline in all cases.

Reviewers: jtolmer, nizamordulu

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D947388

Task ID: 2750307