Really disable perfschema by default.
Summary:
Instead of disabling it in our fb.cmake build config, disable it by
default for all builds.  Perfschma is clearly not supported in 5.6.

Still force it to be enabled for xtrabackup, because that does not
build without it right now (we may fix that in the future).

Internal:
I will squash this with https://phabricator.fb.com/D712554
...but this diff will still be made public in the current branch.

Test Plan:
Built with our tools and plain cmake.
I will do more testing, to be sure, but all looks good.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D947617

Task ID: 2801748