Fix race between semi-sync group commit and dump thread.
Summary:
We changed the bin log dump thread so that it does not need
to acquire LOCK_log. Because of that, it was possible for
the dump thread to read events from the bin log before the
after_flush hook is called. When the dump thread wins that
race, it means that ReplSemiSyncMaster::writeTranxInBinlog
was not yet called and so the dump thread will not request
an ACK from the slave for the group being committed, possibly
resulting in the master waiting for the full semi-sync timeout
to expire before continuing.

Test Plan:
Jenkin's Valgrind builds and
mytest --valgrind --repeat=8 --parallel=8 --big-test\
  rpl.rpl_semi_sync_group_commit_deadlock ...

Internal:
Squash with https://phabricator.fb.com/D781167.

I wrote a test which reproduces the error when the fix is
not in place, but it does not work after the fix is made.
For posterity, the test can be found at
https://phabricator.fb.com/differential/diff/2739915/.

Reviewers: santoshb

Reviewed By: santoshb

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D952336

Task ID: 2566063